requirements.txt
flask
scikit-learn
nltk
chatterbot
sklearn
gunicorn
PyYaml